can run

+++ guided_diffusion/nn.py
@@ -62,7 +62,8 @@ def update_ema(target_params, source_params, rate=0.99):
     :param rate: the EMA rate (closer to 1 means slower).
     """
     for targ, src in zip(target_params, source_params):
-        targ.detach().mul_(rate).add_(src, alpha=1 - rate)
+        # targ.detach().mul_(rate).add_(src, alpha=1 - rate)
+        targ.detach().type_as(src).mul_(rate).add_(src, alpha=1 - rate)
 
 
 def zero_module(module):
